refactor: promote priority election to non-experimental
