+++ qa/integration-tests/src/test/java/io/camunda/zeebe/it/clustering/FixedPartitionDistributionIT.java
@@ -89,8 +89,8 @@ final class FixedPartitionDistributionIT {
     broker.setDockerImageName(
         ZeebeTestContainerDefaults.defaultTestImage().asCanonicalNameString());
     broker
+        .withEnv("ZEEBE_BROKER_CLUSTER_RAFT_ENABLE_PRIORITY_ELECTION", "false")
         .withEnv("ZEEBE_BROKER_EXPERIMENTAL_PARTITIONING_SCHEME", "FIXED")
-        .withEnv("ZEEBE_BROKER_EXPERIMENTAL_ENABLE_PRIORITY_ELECTION", "false")
         .withEnv("ZEEBE_BROKER_EXPERIMENTAL_PARTITIONING_FIXED_0_PARTITIONID", "1")
         .withEnv("ZEEBE_BROKER_EXPERIMENTAL_PARTITIONING_FIXED_0_NODES_0_NODEID", "1")
         .withEnv("ZEEBE_BROKER_EXPERIMENTAL_PARTITIONING_FIXED_0_NODES_1_NODEID", "2")
